+++ js/data/mainData.js
@@ -9,6 +9,7 @@ import { addEvent, pipe } from '../helper/commonFunction.js';
 import { 
     menuAddTemplate, menuDeleteColumnTemplate, menuUpdateTemplate, menuAddColumnTemplate 
 } from '../template.js';
+import { eventToCardEditBtn, hoverEventToCardDeleteBtn } from './button.js';
 
 NodeList.prototype.forEach = Array.prototype.forEach;
 const $columnContent = document.querySelector("#column_holder");
@@ -34,12 +35,12 @@ class Main {
     }
 
     showMainHTML() {
-        $columnContent.innerHTML = this.columns
-        .reduce((runningString, column) => runningString + column.getTemplate(), '');
+        $columnContent.innerHTML = this.columns.reduce(
+            (runningString, column) => runningString + column.getTemplate(), '');
         eventToColumnHeader();
         eventToCardDeleteBtns();
-        this.hoverEventToCardDeleteBtn();
-        this.eventToCardEditBtn();
+        hoverEventToCardDeleteBtn();
+        eventToCardEditBtn();
         eventToColumnBtns();
         dragHandler();
     }
@@ -66,44 +67,6 @@ class Main {
         ($columnMain) => createCardRegisterForm($columnMain)
     )(document.getElementById(columnID));
 
-    eventToCardAdditonCancelBtn = () => pipe(
-        ($cardCancelBtn) => 
-            addEvent($cardCancelBtn, [() => {
-                const column = $cardCancelBtn.closest(".column_main")
-                column.removeChild(column.firstChild);
-            }
-        ])
-    )(document.querySelector("#card_addition_cancel"));
-
-    eventToCardAdditionConfirmBtn = () => pipe(
-        ($confirmBtn) => addEvent($confirmBtn, [
-            () => {
-                const title = document.getElementsByClassName("card_addition_title")[0].value;
-                const main = document.getElementsByClassName("card_addition_text")[0].value;
-
-                if (title === '' || main === '') return;
-
-                const columnID = $cardConfirmBtn.closest(".column").id;
-                this.columns.forEach((column) => {
-                    if (column.name === columnID) {
-                        const newCard = new Card(title, main, "jaehyun cho");
-
-                        if (cardTitle !== '') {
-                            this.replaceCardWithTitle(cardTitle, newCard);
-                            addSidebarBlock(menuUpdateTemplate(column.name, title));
-                        } else {
-                            column.cards.unshift(newCard);
-                            addSidebarBlock(menuAddTemplate(column.name, title));
-                        }
-                        patchMainData();
-                    }
-                });
-
-                this.showMainHTML();
-            }
-        ])
-    )(document.querySelector("#card_addition_confirm"));
-
     deleteColumn = (columnID) => pipe(
         () => this.columns.find((column) => column.name == columnID),
         (column) => {
@@ -119,46 +82,6 @@ class Main {
         (columnId) => deleteColumnData(columnId)
     )();
 
-    addColumnConfirmController = () => pipe(
-        ($columnConfirmBtn) => addEvent($columnConfirmBtn, [
-            () => {
-                const columnName = $columnConfirmBtn.previousElementSibling.value;
-                if (columnName === '') return;
-                const newColumn = new Column(columnName, []);
-                this.columns.push(newColumn);
-                postColumnData(newColumn);
-                addSidebarBlock(menuAddColumnTemplate(columnName));
-                this.showMainHTML();
-            }
-        ])
-    )(document.querySelector(".column_confirm_button"));
-
-    hoverEventToCardDeleteBtn() {
-        const $deleteBtns = document.querySelectorAll(".card_delete_button");
-        $deleteBtns.forEach(($btn) => {
-            const targetCard = $btn.closest(".card");
-
-            addEvent($btn, [
-                () => targetCard.style.outline = "solid #FF4343",
-                () => targetCard.style.backgroundColor = "#FFEEEC"
-            ], "mouseover");
-
-            addEvent($btn, [
-                () => targetCard.style.outline = "none",
-                () => targetCard.style.backgroundColor = "#FFFFFF"
-            ], "mouseout");
-        });
-    }
-
-    eventToCardEditBtn() {
-        const $modifyBtns = document.querySelectorAll(".card_edit_button");
-        $modifyBtns.forEach(($btn) => {
-            addEvent($btn, [() => this.switchAddCardHTML($btn.closest(".card"))]);
-            addEvent($btn, [() => $btn.style.color = "#0075DE"], "mouseover");
-            addEvent($btn, [() => $btn.style.color = "#000000"], "mouseout");
-        });
-    }
-
     switchAddCardHTML = (targetCard) => pipe(
         ($newCard) => {
             targetCard.parentElement.replaceChild($newCard, targetCard);
@@ -168,7 +91,7 @@ class Main {
             addEvent($cancelBtn, [() => 
                 $cardAddCancelBtn.closest(".column_main").replaceChild(targetCard, newCard)]),
         () => {
-            this.eventToCardAdditionConfirmBtn(cardTitle);
+            eventToCardAdditionConfirmBtn(cardTitle);
             resizeCardInput();
         }
     )(getCardRegisterNode(targetCard.id, targetCard.querySelector(".card_main_text").innerHTML));
