+++ js/dragHandler.js
@@ -5,16 +5,17 @@ import { patchMainData } from "./server/mainData.js";
 import { menuMoveTemplate } from "./template.js";
 
 let $dragCard = null;
+let prevColumnID = "";
 HTMLCollection.prototype.forEach = Array.prototype.forEach;
 
 // drag event to card
 const dragStartEventToCard = ($card) => addEvent($card, [
     (event) => $dragCard = event.target,
-    (event) => event.target.classList.add("dragging")
+    ({target}) => target.classList.add("dragging")
 ], "dragstart");
 
 const dragEndEventToCard = ($card) => addEvent($card, [
-    (event) => event.target.classList.remove("dragging")
+    ({target}) => target.classList.remove("dragging")
 ], "dragend")
 
 const dragEventToEveryCard = () => pipe(
@@ -25,19 +26,19 @@ const dragEventToEveryCard = () => pipe(
 )(document.querySelectorAll(".card"));
 
 // drag event to column
-const dragOverEventToColumn = ($column) => addEvent(
-    $column, [(event) => event.preventDefault()], "dragover");
+const dragOverEventToColumn = ($column) => addEvent($column, 
+    [(event) => event.preventDefault()], "dragover");
 
 const dropEventToColumn = ($column) => addEvent($column, [
-    ({target}) => {
-        if(target.className !== "column") return;
-
-        const prevColumnID = $dragCard.closest("li.column").getAttribute("id");
+    () => { if(target.className !== "column") return;},
+    () => {
         $dragCard.remove();
-
+        target.appendChild($dragCard);
+    },
+    () => prevColumnID = $dragCard.closest("li.column").getAttribute("id"),
+    ({target}) => {
         const prevColumn = main.columns.filter((column) => prevColumnID == column.getName())[0];
         const nextColumn = main.columns.filter((column) => column.name == target.id)[0];
-
         for(const [index, card] of prevColumn.cards.entries()) {
             if(card.title !== $dragCard.id) continue;
             prevColumn.cards.splice(index, 1);
@@ -45,20 +46,17 @@ const dropEventToColumn = ($column) => addEvent($column, [
         }
 
         addSidebarBlock(menuMoveTemplate($dragCard.id, prevColumn.getName(), nextColumn.name));
-
-        target.appendChild($dragCard);
     },
     () => patchMainData(),
     () => main.updateColumnLength()
 ], "drop");
 
-const dragEventToEveryColumn = () => {
-    const columnArray = document.querySelectorAll(".column");
-    columnArray.forEach(($column) => {
+const dragEventToEveryColumn = () => pipe(
+    (columnArray) => columnArray.forEach(($column) => {
         dragOverEventToColumn($column);
         dropEventToColumn($column);
     })
-}
+)(document.querySelectorAll(".column"));
 
 const dragHandler = () => pipe(
     dragEventToEveryCard,
